pricing_service/requirements.txt
aioredis==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
fastapi-cache==0.1.0
fastapi-cache2==0.2.1
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pendulum==3.0.0
pika==1.3.2
pluggy==1.4.0
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.4
pytest-asyncio==0.23.4
python-dateutil==2.8.2
python-dotenv==1.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
time-machine==2.13.0
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.27.0
